+++ gomill/competitions.py
@@ -50,6 +50,7 @@ _player_settings = [
     Setting('startup_gtp_commands', allow_none(interpret_sequence),
             defaultmaker=list),
     Setting('discard_stderr', interpret_bool, default=False),
+    Setting('sgf_player_name_from_gtp', interpret_bool, default=True),
     ]
 
 class Player_config(Quiet_config):
@@ -252,6 +253,8 @@ class Competition(object):
 
         player.is_reliable_scorer = config['is_reliable_scorer']
         player.allow_claim = config['allow_claim']
+        player.discard_stderr = config['discard_stderr']
+        player.sgf_player_name_from_gtp = config['sgf_player_name_from_gtp']
 
         player.startup_gtp_commands = []
         try:
@@ -280,9 +283,6 @@ class Competition(object):
                 player.gtp_aliases[cmd1] = cmd2
         except ValueError, e:
             raise ControlFileError("'gtp_aliases': %s" % e)
-
-        if config['discard_stderr']:
-            player.discard_stderr = True
 
         return player